Add support for changeable cell stratification type (#26)
* Add changeable cell stratification type

* clean up
